Dungeon Images
Updated the code for the dungeon (and heroic) images behind the archmages in Dalaran. They are now essentially stunned and all of them do actually spin around as they should. It's not perfect but it works great.

Some of them appear to be hovering, or levitating, but this is actually as it should be. I compared it to pictures I found on wowhead and old videos from wrath of the lich king.