Patches
Preparing to add support for all patches that my core patch includes. I will be porting the data over but still focus on Wrath of the Lich King. Right now the data is the same as before so there's no real difference aside from a few config options and damage/healing values yet.