Argent Tournament Grounds
Fix an issue where players using flight paths would get teleported out of the argent tournament grounds and interrupt their flight, and essentially forcing them to travel elsewhere to continue.

There is an obvious drawback to this. If a player for whatever reason has the flight path for the grounds unlocked, they can enter it. It is for now not too important since no one should be able to unlock it.