DBC
Added DBCs that I use to play the game. I don't use the talent trees right now but when I do I use these based on the expansion. The talent trees needs to be both server- and client-side but Onyxia's Lair only needs to be client-side, and all it does is show the 40-player icon on the minimap.

When the level 60 Naxxramas is implemented I will create folders showing what patches each file should be used for.