DBCs
Updated the DBCs based on patch range. There are more changes to be done but the current setup is the base for my work towards vanilla Naxxramas.

I haven't yet moved on to TBC and won't for quite some time so the TBC talents aren't done.

The talents are NOT restored vanilla talent trees. That would require a lot of spell changes and I won't be doing that. It's wrath talents with a lot of the talents removed to make it look like the vanilla trees.

The .adt/.wdt files is to change Eastern Plaguelands and restore the original Naxxramas entrance. I'm working on the SQL for it on the side and won't be adding it here until I'm satisfied, it is a ton of work after all.