Several changes and additions
Added events for the AQ gate and the gong.
Changed the init process to be executed in OnBeforeWorldInitialized rather than OnStartup. Preloading maps messes with stuff and this fixes that.
Changed the ID of all events. Trying to use an ID above 127 causes serious issues and would probably need a core change.