Added the final spawn points for the post-opening event
There are a few invisible creatures to add that handle the mind control part of these crystals, but after that this part of the event should be done.

The smart scripts for most of these are guesstimates in terms of timers and in some cases spell ids. It needs a lot of testing.