Cleaned up playerbots talents
The talents are the same but I removed unnecessary spaces and added the PvP specs. It's important to note that the PvP specs are exactly the same as the PvE one. This means that if a bot decides to use a PvP spec they won't be using the original talent trees.

The talent trees for death knight are untouched. I won't be changing those talent trees for vanilla or TBC.

I am slowly working my way through the TBC talent trees and will be adding them to this repository when I'm done.